+++ NEWS.md
@@ -1,3 +1,11 @@
+# `CAST` 0.7.0
+* new feature: 
+  * nndm cross-validation as suggested by Milà et al. (2022)
+* modifications
+  * plot_geodist works with NNDM
+  * trainDI works with NNDM
+  * rename of parameter folds in AOA and trainDI
+
 # `CAST` 0.6.0
 * new feature: 
   * trainDI allows to calculate the DI of the training dataset separately from the aoa function